fix: include remaining local updates

+++ lib/features/activity/application/activity_time_sync_engine.dart
@@ -450,7 +450,7 @@ class ActivityTimeSyncEngine {
     final int clamped = rawMinutes.clamp(
       _earliestCandidateMinutes,
       _latestCandidateMinutes,
-    ) as int;
+    );
     return ((clamped / 5).round()) * 5;
   }
